Initial support for DNS timeouts
