+++ src/platform.h
@@ -50,24 +50,25 @@ int real_main(int argc, char *argv[]);
 
 #endif // CONFIG_SDL
 
-// Normally only need mutexes for audio, but the Wii port
+// Need threads and mutexes for DNS lookups.
+// Otherwise only need mutexes for audio, but the Wii port
 // uses them for events too
 
-#if defined(CONFIG_WII) || defined(CONFIG_AUDIO)
+#if defined(CONFIG_AUDIO) || defined(CONFIG_NETWORK) || defined(CONFIG_WII)
 
-#ifdef CONFIG_PTHREAD_MUTEXES
-#include "mutex_pthread.h"
+#ifdef CONFIG_PTHREAD
+#include "thread_pthread.h"
 #elif defined(CONFIG_SDL)
-#include "mutex_sdl.h"
+#include "thread_sdl.h"
 #elif defined(CONFIG_WII)
 #include "../arch/wii/mutex.h"
 #elif defined(CONFIG_3DS)
 #include "../arch/3ds/mutex.h"
 #else
-#error Provide a valid mutex implementation for this platform!
+#error Provide a valid thread/mutex implementation for this platform!
 #endif
 
-#endif // defined(CONFIG_WII) || defined(CONFIG_AUDIO)
+#endif // defined(CONFIG_AUDIO) || defined(CONFIG_NETWORK) || defined(CONFIG_WII)
 
 CORE_LIBSPEC void delay(Uint32 ms);
 CORE_LIBSPEC Uint32 get_ticks(void);
